Remove `_report` package.
