+++ src/gpsea/view/__init__.py
@@ -4,7 +4,7 @@ from ._disease import DiseaseViewer
 from ._phenotype_analysis import summarize_hpo_analysis
 from ._protein_viewer import ProteinVariantViewer
 from ._protein_visualizable import ProteinVisualizable
-from ._report import GpseaReport
+from ._base import GpseaReport
 from ._stats import MtcStatsViewer
 from ._txp import VariantTranscriptVisualizer
 from ._protein_visualizer import ProteinVisualizer
